fix: refactor packages
