+++ backend/src/main/java/com/exadel/carinsurance/model/request/AuthRequestEntity.java
@@ -1,4 +1,4 @@
-package com.exadel.carinsurance.model.response;
+package com.exadel.carinsurance.model.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,9 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class RegisterRequestEntity {
-  private String firstName;
-  private String lastName;
+public class AuthRequestEntity {
   private String email;
   private String password;
 }
