+++ backend/src/main/java/com/exadel/carinsurance/model/request/RegisterRequestEntity.java
@@ -1,4 +1,4 @@
-package com.exadel.carinsurance.model.response;
+package com.exadel.carinsurance.model.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class AuthRequestEntity {
+public class RegisterRequestEntity {
+  private String firstName;
+  private String lastName;
   private String email;
   private String password;
 }
