+++ backend/src/main/java/com/exadel/carinsurance/service/implementation/AuthService.java
@@ -6,9 +6,9 @@ import com.exadel.carinsurance.mapper.UserResponseMapper;
 import com.exadel.carinsurance.model.ERoleEntity;
 import com.exadel.carinsurance.model.RoleEntity;
 import com.exadel.carinsurance.model.UserEntity;
-import com.exadel.carinsurance.model.response.AuthRequestEntity;
+import com.exadel.carinsurance.model.request.AuthRequestEntity;
+import com.exadel.carinsurance.model.request.RegisterRequestEntity;
 import com.exadel.carinsurance.model.response.AuthResponseEntity;
-import com.exadel.carinsurance.model.response.RegisterRequestEntity;
 import com.exadel.carinsurance.model.response.UserResponseEntity;
 import com.exadel.carinsurance.repository.IRoleRepository;
 import com.exadel.carinsurance.repository.IUserRepository;
